[docs] Add summary table for hardware requirements and more detailed on storage
Took the same requirement as described in https://cozystack.io/docs/getting-started/first-deployment/.

Added the primary disk description to clearly mention that Cozystack use it as the base system storage. Also workload data are sotre using the StorageClass defined once Linstor is set up. (Added relevant source for storage configuration) https://cozystack.io/docs/getting-started/first-deployment/#configure-storage

Also added a little explanation of a common issue encountered in the deployment lifecycle where the system disk is not /dev/sda as Talos bootsrap methods might not find system disk and refuse to install. Pointed to the issue cozystack/cozystack#723 where the user failed to botostrap with talm likely because of the storage not defined as /dev/sda

Signed-off-by: lb0o <lb0o@nocilo.com>